django/requirements.txt
Django==4.1.4
djangorestframework==3.14.0
drf-social-oauth2==1.2.1
django-cors-headers==3.13.0
sqlparse==0.4.3
mysqlclient==2.1.1
pytz==2022.6
uWSGI==2.0.21